Replaced devtools components with devtools extension
