+++ package.json
@@ -34,9 +34,6 @@
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.2",
     "redux": "^3.5.2",
-    "redux-devtools": "^3.2.0",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.1.1",
     "redux-immutable": "^3.0.6",
     "reselect": "^2.5.1"
   },
@@ -52,9 +49,6 @@
     "@types/react-redux": "^4.4.32",
     "@types/react-router": "^3.0.0",
     "@types/react-router-redux": "^4.0.33",
-    "@types/redux-devtools": "^3.0.33",
-    "@types/redux-devtools-dock-monitor": "^1.1.29",
-    "@types/redux-devtools-log-monitor": "^1.0.29",
     "@types/redux-immutable": "^3.0.31",
     "@types/redux-logger": "^2.6.32",
     "browserify": "^14.0.0",
