--- src/containers/dev_tools.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import * as ReduxDevtools from "redux-devtools";
-import LogMonitor from "redux-devtools-log-monitor";
-import DockMonitor from "redux-devtools-dock-monitor";
-
-let DevTools: any = ReduxDevtools.createDevTools(
-    <DockMonitor toggleVisibilityKey="ctrl-h" changePositionKey="ctrl-w">
-        <LogMonitor />
-    </DockMonitor>
-);
-
-export default DevTools;
