--- src/store/configure_store.prod.ts
@@ -1,18 +1,0 @@
-import { createStore, applyMiddleware } from "redux";
-import * as Redux from "redux";
-
-function configureStore(
-  middlewares: Redux.Middleware[],
-  rootReducer: Redux.Reducer<any>,
-  initialState: any
-): Redux.Store<any> {
-
-  return createStore(
-    rootReducer,
-    initialState,
-    applyMiddleware(...middlewares)
-  );
-
-}
-
-export default configureStore;
